feat: database adjust + health check fixed
environment variable (database) fixed, health check from backend and
front commente, and from database fixed
