--- backend/auth_app/static/css/base.css
@@ -1,1 +1,0 @@
-/* for general styles (fonts, colors, resets) */
