--- backend/auth_app/static/css/component.css
@@ -1,4 +1,0 @@
-/*/*  Reusable UI components (buttons, cards) */*/
-/*.signin{*/
-
-}
