--- backend/auth_app/static/css/layout.css
@@ -1,1 +1,0 @@
-/* Layout related css (grid, flexbox) */
