--- backend/auth_app/static/css/styles.css
@@ -1,67 +1,0 @@
-body, html {
-    margin: 0;
-    padding: 0;
-    height: 100%;
-    background-color: #000808;
-    color: #00ff99;
-    font-family: 'Press Start 2P', cursive;
-    text-align: center;
-    overflow: hidden;
-}
-
-body {
-    background-image: linear-gradient(to right, rgba(0, 255, 153, 0.1) 1px, transparent 1px),
-                      linear-gradient(to bottom, rgba(0, 255, 153, 0.1) 1px, transparent 1px );
-    background-size: 40px 40px;
-}
-
-.crt-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: repeating-linear-gradient(
-        rgba(0, 0, 0, 0.1),
-        rgba(0, 0, 0, 0.1) 1px,
-        transparent 1px,
-        transparent 2px
-    );
-    pointer-events: none;
-    z-index: 1;
-}
-
-.title {
-    margin-top: 10%;
-    font-size: 1.5rem;
-    text-shadow: 0 0 3px #00ff99, 0 0 10px #00ff99;
-}
-
-.menu {
-    margin-top: 2rem;
-}
-
-button, .star-us {
-    display: block;
-    width: 300px;
-    margin: 10px auto;
-    padding: 10px;
-    color: #00ff99;
-    background-color: #011c1a;
-    border: 2px solid #00ff99;
-    font-size: 1rem;
-    cursor: pointer;
-    text-decoration: none;
-    box-shadow: 0 0 8px #00ff99, 0 0 22px #00ff99;
-    transition: all 0.2s ease-in-out;
-}
-
-button:hover, .star-us:hover {
-    background-color: #00ff99;
-    color: #000808;
-    box-shadow: 0 0 10px #00ff99, 0 0 40px #00ff99
-}
-
-.star-us{
-    font-size: 0.8rem;
-}
