--- backend/auth_app/templates/home.html
@@ -1,1 +1,0 @@
-<p>fucker</p>
