--- backend/auth_app/templates/login.html
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-
-<a href="/auth/42">
-	<button class="sign-in">Sign in with 42</button>
-</a>
-
