--- backend/auth_app/templates/logout.html
@@ -1,9 +1,0 @@
-{% if messages %}
-	<div>
-		{% for messages in messages %}
-			<div class="alert alert-{{ messages.tags}}">
-				{{ messages }}
-			</div>
-		{% endfor %}
-	</div>
-{% endif %}
