+++ backend/transcendence/static/css/base.css
@@ -1,0 +1,1 @@
+/* for general styles (fonts, colors, resets) */
