+++ backend/transcendence/static/css/layout.css
@@ -1,0 +1,1 @@
+/* Layout related css (grid, flexbox) */
