+++ backend/transcendence/templates/home.html
@@ -1,0 +1,1 @@
+<p>fucker</p>
