+++ backend/transcendence/templates/login.html
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+
+<a href="/auth/42">
+	<button class="sign-in">Sign in with 42</button>
+</a>
+
