+++ backend/transcendence/templates/logout.html
@@ -1,0 +1,9 @@
+{% if messages %}
+	<div>
+		{% for messages in messages %}
+			<div class="alert alert-{{ messages.tags}}">
+				{{ messages }}
+			</div>
+		{% endfor %}
+	</div>
+{% endif %}
