+++ backend/transcendence/urls.py
@@ -1,0 +1,32 @@
+"""
+URL configuration for transcendence project.
+
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/5.1/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.urls import include, path
+    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+"""
+from django.contrib import admin
+from django.urls import path
+from django.conf import settings
+from django.conf.urls.static import static
+
+from user_auth import requestAuth, logout, home, dashboard
+
+
+urlpatterns = [
+    path('', home.home, name='home'),
+    path('admin/', admin.site.urls),
+    path('auth/42/', requestAuth.requestAuth42, name='requestAuth42'),
+    path('auth/42/callback/', requestAuth.callbackAuth, name='callbackAuth'),
+    path('logout/', logout.logout, name='logout'),
+    path('dashboard/', dashboard.dashboard, name='dashboard')
+] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
