+++ docker-compose.yml
@@ -17,16 +17,19 @@ services:
     env_file:
       - .env
     environment:
-      - "POSTGRES_DB=postgres"
-      - "POSTGRES_USER=postgres"
-      - "POSTGRES_PASSWORD=postgres"
+      - "DATABASE=postgres"
+      - "DB_NAME=postgres"
+      - "DB_USER=postgres"
+      - "DB_PASSWORD=postgres"
+      - "DB_HOST=db"
+      - "DB_PORT=5432"
     volumes:
       - db-volume:/var/lib/postgresql/data/
     restart: always
     ports:
      - "5432:5432"
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready"]
+      test: ["CMD-SHELL", "pg_isready -U postgres -d postgres -h localhost"]
       interval: 1m30s
       timeout: 30s
       retries: 5
